break out options to a separate file
