+++ Makefile.am
@@ -4,7 +4,7 @@
 AUTOMAKE_OPTIONS = gnu # I would like dist-bzip2 here, but automake complains
 bin_PROGRAMS = rdfind
 rdfind_SOURCES = rdfind.cc Checksum.cc  Dirlist.cc  Fileinfo.cc  Rdutil.cc \
-                 EasyRandom.cc UndoableUnlink.cc CmdlineParser.cc
+                 EasyRandom.cc UndoableUnlink.cc CmdlineParser.cc Options.cc
 
 LDADD = @LIBXXHASH@
 #these are the test scripts to execute - I do not know how to glob here,
@@ -41,7 +41,7 @@ AUXFILES=testcases/common_funcs.sh \
 EXTRA_DIST = \
   Dirlist.hh Checksum.hh  Fileinfo.hh \
   Rdutil.hh bootstrap.sh RdfindDebug.hh EasyRandom.hh UndoableUnlink.hh \
-  CmdlineParser.hh \
+  CmdlineParser.hh Options.hh \
   $(TESTS) \
   $(AUXFILES) \
   rdfind.1 LICENSE \
